# [WIP] Dictionary - magic remainder