Split Resize -> Expand and Rehash

Rehashing is only used in `string` key variant so it doesn't need to be compiled in every Expand
Rehashing doesn't expand the Dictionary so it can reuse the existing arrays
Use MagicNumberRemainder rather than `idiv` in `Expand` and `Rehash`