@benaadams 2017-12-09
Before `Resize` 
```asm
; Lcl frame size = 88
...
       33DB                 xor      ebx, ebx
       8944244C             mov      dword ptr [rsp+4CH], eax
       85C0                 test     eax, eax
       7E4C                 jle      SHORT G_M29783_IG12
       458B4508             mov      r8d, dword ptr [r13+8]

G_M29783_IG10:
       413BD8               cmp      ebx, r8d
       0F8393000000         jae      G_M29783_IG18
       4863D3               movsxd   rdx, ebx
       488D1452             lea      rdx, [rdx+2*rdx]
       498D4CD510           lea      rcx, bword ptr [r13+8*rdx+16]
       448B4910             mov      r9d, dword ptr [rcx+16]
       4585C9               test     r9d, r9d
       7C1C                 jl       SHORT G_M29783_IG11
       418BC1               mov      eax, r9d
       99                   cdq      
       F7FF                 idiv     edx:eax, edi
       413BD6               cmp      edx, r14d
       7373                 jae      SHORT G_M29783_IG18
       4863C2               movsxd   rax, edx
       8B448510             mov      eax, dword ptr [rbp+4*rax+16]
       894114               mov      dword ptr [rcx+20], eax
       4863C2               movsxd   rax, edx
       895C8510             mov      dword ptr [rbp+4*rax+16], ebx

G_M29783_IG11:
       FFC3                 inc      ebx
       8B44244C             mov      eax, dword ptr [rsp+4CH]
       3BD8                 cmp      ebx, eax
       8944244C             mov      dword ptr [rsp+4CH], eax
       7CB8                 jl       SHORT G_M29783_IG10
...
; Total bytes of code 497, prolog size 29 for method Dictionary`2:Resize(int,bool):this
```

After `Expand`
```asm
; Lcl frame size = 72
...
       33C0                 xor      eax, eax
       85F6                 test     esi, esi
       7E58                 jle      SHORT G_M13446_IG08
       418B5508             mov      edx, dword ptr [r13+8]

G_M13446_IG06:
       3BC2                 cmp      eax, edx
       737A                 jae      SHORT G_M13446_IG10
       4863C8               movsxd   rcx, eax
       488D0C49             lea      rcx, [rcx+2*rcx]
       4D8D44CD10           lea      r8, bword ptr [r13+8*rcx+16]
       458B4810             mov      r9d, dword ptr [r8+16]
       4585C9               test     r9d, r9d
       7C35                 jl       SHORT G_M13446_IG07
       418BC9               mov      ecx, r9d
       448BD5               mov      r10d, ebp
       4C0FAFD1             imul     r10, rcx
       418D4E20             lea      ecx, [r14+32]
       49D3EA               shr      r10, cl
       498BCA               mov      rcx, r10
       8BC9                 mov      ecx, ecx
       0FAFCB               imul     ecx, ebx
       442BC9               sub      r9d, ecx
       453BCC               cmp      r9d, r12d
       7344                 jae      SHORT G_M13446_IG10
       4963C9               movsxd   rcx, r9d
       418B4C8F10           mov      ecx, dword ptr [r15+4*rcx+16]
       41894814             mov      dword ptr [r8+20], ecx
       4963C9               movsxd   rcx, r9d
       4189448F10           mov      dword ptr [r15+4*rcx+16], eax

G_M13446_IG07:
       FFC0                 inc      eax
       3BC6                 cmp      eax, esi
       7CAC                 jl       SHORT G_M13446_IG06
...
; Total bytes of code 316, prolog size 26 for method Dictionary`2:Expand(int):this
```