@benaadams 2017-12-09
I think the splitting of Resize stands alone as a good change so will add a PR for that.