@benaadams 2017-12-19
Last commit https://github.com/dotnet/coreclr/pull/15453/commits/a8917c0b5fbb8c4be68f3782a13bd7fcec60cc4b adds magic remainder onto https://github.com/dotnet/coreclr/pull/15419

Its definitely faster 5% - 10%

```
                           Method |   Items |            Current |                New |  Change |
--------------------------------- |-------- |-------------------:|-------------------:|--------:|
    ContainsKey_Single_Yes_IntKey |     100 |           5.909 ns |           5.629 ns |  x 1.04 |
     ContainsKey_Single_No_IntKey |     100 |           5.973 ns |           5.681 ns |  x 1.05 |
           ContainsKey_All_IntKey |     100 |         591.147 ns |         580.291 ns |  x 1.02 |
 ContainsKey_Single_Yes_StringKey |     100 |          20.476 ns |          18.509 ns |  x 1.10 |
  ContainsKey_Single_No_StringKey |     100 |          16.165 ns |          14.430 ns |  x 1.12 |
 ContainsKey_Single_Yes_ObjectKey |     100 |           8.749 ns |           7.798 ns |  x 1.12 |
  ContainsKey_Single_No_ObjectKey |     100 |           8.448 ns |           7.737 ns |  x 1.09 |
                       Add_IntKey |     100 |       3,989.325 ns |       3,565.964 ns |  x 1.11 |
                    Add_StringKey |     100 |       9,967.582 ns |       9,539.857 ns |  x 1.04 |
                    Add_ObjectKey |     100 |       8,678.552 ns |       8,191.180 ns |  x 1.06 |
               PreSizedAdd_IntKey |     100 |       1,105.320 ns |       1,026.360 ns |  x 1.08 |
            PreSizedAdd_StringKey |     100 |       7,460.880 ns |       7,082.505 ns |  x 1.05 |
            PreSizedAdd_ObjectKey |     100 |       6,240.259 ns |       6,046.344 ns |  x 1.03 |
```
However, there is a lot going on, including extra memory accesses to get the magic and shift (and makes the dictionary 2 fields larger)

Changes
```asm
and      ebx, 0xD1FFAB1E           ;  & 0x7FFFFFFF
mov      rax, gword ptr [rsi+8]    
mov      rcx, rax                  
mov      r8d, dword ptr [rax+8]    
mov      eax, ebx                  
cdq                                
idiv     edx:eax, r8d              
cmp      edx, r8d                  
jae      G_M6157_IG09              ; CORINFO_HELP_RNGCHKFAIL
```
to
```asm
and      r15d, 0xD1FFAB1E          ;  & 0x7FFFFFFF
mov      edx, dword ptr [rbx+8]    
mov      r8d, edx                  
mov      ecx, dword ptr [rdi+72]   
mov      r11d, dword ptr [rdi+76]  
mov      eax, r15d                 
mov      ecx, ecx                  
imul     rax, rcx                  
mov      ecx, r11d                 
shr      rax, cl                   
mov      rcx, rax                  
mov      ecx, ecx                  
imul     ecx, r8d                  
mov      r8d, r15d                 
sub      r8d, ecx                  
cmp      r8d, edx                  
jae      SHORT G_M6159_IG07        ; CORINFO_HELP_RNGCHKFAIL
```
So there's probably a better way; for more gain with less cost. However, its here as an investigation for posterity to be referred to.