@mikedn 2017-12-19
> Its definitely faster 5% - 10%

Hmm, not as much as I would have hopped. But then you replace `div` with a memory load, 2 imuls, 1 shift and one subtract - that's something like 15 cycles instead of 26. Not bad but not a very big difference either.

> mov      ecx, ecx                  ; bit weird

Those are likely `uint`->`long` casts. Well, the first certainly is, not so sure about the second, it seems pointless since the `imul` that follows is 32 bit, not 64.

> not sure the asm is great?

Yeah, on CPU that support it `SHRX` could be used instead of `SHR`. Should save 2 cycles on Intel CPUs.

> So there's probably a better way; for more gain with less cost

Well, the only other solution is to completely get rid of the need for prime numbers and use power of 2 numbers instead... and deal with all the baggage that such a change would bring.