@benaadams 2017-12-19
I tried the [fast range](https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/) and while it was much faster; the benchmarker (in-proc) started crawling after a while (while collating results) - so they are definitely badly distributed hashes around.

Thought about trying rehashing the hash with `Sse42.Crc32` but that's not available everywhere 😢